Create Fashion-MNIST1
Fashion-MNIST is a dataset of Zalando's (http://www.zalando.com) article images —consisting of a training set of 60,000 examples and a test set of 10,000 examples. Each example is a 28x28 grayscale image, associated with a label from 10 classes. Fashion-MNIST serves as a direct drop-in replacement for the original MNIST dataset for benchmarking machine learning algorithms. It shares the same image size and structure of training and testing splits. (See GitHub Repo)

The objective of the project is to use Fashion-MNIST data set to identify the different fashion products from the given pictures using various best possible models (ML algorithms) and report the values of the performance measures for different models. Also, report the model that performs best, and fine-tune the same model using one of the model fine-tuning techniques, and report the best possible combination of hyperparameters for the selected model. Lastly, use the selected model to make final predictions and report the values of various performance measures for the same.

Hint: You can use dimensionality reduction to simplify the things.

filePath = '/cxldata/datasets/project/fashion-mnist/’